requirements.txt
flask
psycopg2-binary
PyJWT
python-dotenv
werkzeug
